Move resource metadata to content and fix population hover
